Restructure packaging for seven eleven
COAND-497
